Seguridad en rutas

+++ src/controllers/photo.controllers.ts
@@ -2,11 +2,16 @@ import {Request,Response} from 'express';
 import Photo from '../models/photo';
 import path from 'path';
 import fs from 'fs-extra';
-
+import admin from '../models/admin';
 
 
 export async function createPhoto (req:Request, res:Response){
- 
+    
+    const user = await admin.findById(req.userId, { password: 0 });
+    if (!user) {
+        return res.status(404).json('No User found');
+    }
+
     const {title,tag,description} = req.body;
 
     const newPhoto = {
@@ -37,18 +42,34 @@ export async function createPhoto (req:Request, res:Response){
 }
 
 export async function getPhotos (req:Request, res:Response): Promise <Response> {
+    
+    const user = await admin.findById(req.userId, { password: 0 });
+    if (!user) {
+        return res.status(404).json('No User found');
+    }
+
     const photos = await Photo.find();
     return res.json(photos);
 
 }
 
 export async function getPhoto (req:Request, res:Response): Promise <Response> {
+    const user = await admin.findById(req.userId, { password: 0 });
+    if (!user) {
+        return res.status(404).json('No User found');
+    }
+
     const photo = await Photo.findById(req.params.id);
     return res.json(photo);
 
 }
 
 export async function DeletePhoto(req:Request, res:Response):Promise <Response>{
+
+    const user = await admin.findById(req.userId, { password: 0 });
+    if (!user) {
+        return res.status(404).json('No User found');
+    }
 
     const photo = await Photo.findByIdAndRemove(req.params.id);
 
@@ -63,6 +84,11 @@ export async function DeletePhoto(req:Request, res:Response):Promise <Response>{
 }
 
 export async function updatePhoto(req:Request, res:Response):Promise <Response> {
+    const user = await admin.findById(req.userId, { password: 0 });
+    if (!user) {
+        return res.status(404).json('No User found');
+    }
+
     const {title,description}=req.body;
     const updatedPhoto = await Photo.findByIdAndUpdate(req.params.id, {
         title,
